pq-rpm: bigger refactor of the code
Support '%(upstreamversion)s' as a replacable string in the pq branch
name.

Eliminate some unneeded branch switches that were only needed for
finding/parsing the spec file. Now, spec file is directly searched from
the git repository (from the 'base' branch) if the command is run when
on the pq branch - except for the 'export' subcommand which always
switches to the base branch.

Change-Id: Id025a1c404e14f4a16dea7feacc3d352ee19ccb8
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>