pq-rpm: implement '--import-files' command line option
For defining the packaging file(s) that will be imported into the
development/patch-queue branch.

By default, the local gbp conf files are imported.

NOTE: This option does not affect the patch files that are imported. The
files defined with this option will appear as one monolithic commit in
the development/patch-queue branch.

Change-Id: Ia1967d77d2be5a8afc41e8d32c84b3386bc45924
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>