pq-rpm: implement --retain-history option
With this option defined gbp tries to preserve the history when
converting. That is, for each commit in the old branch create one
corresponding commit in the new orphan packaging branch. This works by
dumping packaging files and updating patches for each commit. However,
empty commits are not generated - these are caused e.g. by changes in
files that are ignored by patch-generation.

NOTE! Only valid for the 'convert' action.

Change-Id: I40f6391b2b8f7efd772d138cbb8959f1ab9cf8e0
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>