buildpackage-rpm: always create tarball for native packages
If the package is native, always (re-)create the source tarball, even if
--git-no-create-orig is used. Not creating the tarball does not make
much sense (and most probably just causes problems and confusion): for
native packages there is no place for external tarball and no patches so
we need to always have the correct sources (i.e. the exported revision)
in the tarball.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>